+++ lib/packages/wikibase/data-model-serialization/src/DeserializerFactory.php
@@ -5,10 +5,9 @@ namespace Wikibase\DataModel;
 use Deserializers\Deserializer;
 use Deserializers\DispatchableDeserializer;
 use Deserializers\DispatchingDeserializer;
-use Wikibase\DataModel\Deserializers\AliasGroupDeserializer;
+use Wikibase\DataModel\Deserializers\AliasGroupListDeserializer;
 use Wikibase\DataModel\Deserializers\ClaimsDeserializer;
 use Wikibase\DataModel\Deserializers\EntityIdDeserializer;
-use Wikibase\DataModel\Deserializers\FingerprintDeserializer;
 use Wikibase\DataModel\Deserializers\ItemDeserializer;
 use Wikibase\DataModel\Deserializers\PropertyDeserializer;
 use Wikibase\DataModel\Deserializers\ReferenceDeserializer;
@@ -29,6 +28,7 @@ use Wikibase\DataModel\Entity\EntityIdParser;
  *
  * @licence GNU GPL v2+
  * @author Thomas Pellissier Tanon
+ * @author Bene* < benestar.wikimedia@gmail.com >
  */
 class DeserializerFactory {
 
@@ -57,10 +57,20 @@ class DeserializerFactory {
 	 * @return DispatchableDeserializer
 	 */
 	public function newEntityDeserializer() {
-		$fingerprintDeserializer = new FingerprintDeserializer();
 		return new DispatchingDeserializer( array(
-			new ItemDeserializer( $this->newEntityIdDeserializer(), $fingerprintDeserializer, $this->newStatementListDeserializer(), $this->newSiteLinkDeserializer() ),
-			new PropertyDeserializer( $this->newEntityIdDeserializer(), $fingerprintDeserializer, $this->newStatementListDeserializer() )
+			new ItemDeserializer(
+				$this->newEntityIdDeserializer(),
+				$this->newTermListDeserializer(),
+				$this->newAliasGroupListDeserializer(),
+				$this->newStatementListDeserializer(),
+				$this->newSiteLinkDeserializer()
+			),
+			new PropertyDeserializer(
+				$this->newEntityIdDeserializer(),
+				$this->newTermListDeserializer(),
+				$this->newAliasGroupListDeserializer(),
+				$this->newStatementListDeserializer()
+			)
 		) );
 	}
 
@@ -180,7 +190,7 @@ class DeserializerFactory {
 	}
 
 	/**
-	 * Returns a Deserializer that can serialize Term objects.
+	 * Returns a Deserializer that can deserialize Term objects.
 	 *
 	 * @since 1.5
 	 *
@@ -191,7 +201,7 @@ class DeserializerFactory {
 	}
 
 	/**
-	 * Returns a Deserializer that can serialize TermList objects.
+	 * Returns a Deserializer that can deserialize TermList objects.
 	 *
 	 * @since 1.5
 	 *
@@ -202,14 +212,14 @@ class DeserializerFactory {
 	}
 
 	/**
-	 * Returns a Deserializer that can serialize AliasGroup objects.
+	 * Returns a Deserializer that can deserialize AliasGroupList objects.
 	 *
 	 * @since 1.5
 	 *
 	 * @return Deserializer
 	 */
-	public function newAliasGroupDeserializer() {
-		return new AliasGroupDeserializer();
+	public function newAliasGroupListDeserializer() {
+		return new AliasGroupListDeserializer();
 	}
 
 }