+++ lib/packages/wikibase/data-model-serialization/tests/unit/Serializers/PropertySerializerTest.php
@@ -4,20 +4,48 @@ namespace Tests\Wikibase\DataModel\Serializers;
 
 use Wikibase\DataModel\Entity\Item;
 use Wikibase\DataModel\Entity\Property;
-use Wikibase\DataModel\Serializers\FingerprintSerializer;
 use Wikibase\DataModel\Serializers\PropertySerializer;
 use Wikibase\DataModel\Snak\PropertyNoValueSnak;
 use Wikibase\DataModel\Statement\StatementList;
+use Wikibase\DataModel\Term\AliasGroupList;
+use Wikibase\DataModel\Term\TermList;
 
 /**
  * @covers Wikibase\DataModel\Serializers\PropertySerializer
  *
  * @licence GNU GPL v2+
  * @author Thomas Pellissier Tanon
+ * @author Bene* < benestar.wikimedia@gmail.com >
  */
 class PropertySerializerTest extends SerializerBaseTest {
 
-	protected function buildSerializer() {
+	protected function buildSerializer( $useObjectsForMaps = false ) {
+		$termListSerializerMock = $this->getMock( '\Serializers\Serializer' );
+		$termListSerializerMock->expects( $this->any() )
+			->method( 'serialize' )
+			->will( $this->returnCallback( function( TermList $termList ) {
+				if ( $termList->isEmpty() ) {
+					return array();
+				}
+
+				return array(
+					'en' => array( 'lang' => 'en', 'value' => 'foo' )
+				);
+			} ) );
+
+		$aliasGroupListSerializerMock = $this->getMock( '\Serializers\Serializer' );
+		$aliasGroupListSerializerMock->expects( $this->any() )
+			->method( 'serialize' )
+			->will( $this->returnCallback( function( AliasGroupList $aliasGroupList ) {
+				if ( $aliasGroupList->isEmpty() ) {
+					return array();
+				}
+
+				return array(
+					'en' => array( 'lang' => 'en', 'values' => array( 'foo', 'bar' ) )
+				);
+			} ) );
+
 		$statementListSerializerMock = $this->getMock( 'Serializers\Serializer' );
 		$statementListSerializerMock->expects( $this->any() )
 			->method( 'serialize' )
@@ -40,9 +68,12 @@ class PropertySerializerTest extends SerializerBaseTest {
 				);
 			} ) );
 
-		$fingerprintSerializer = new FingerprintSerializer( FingerprintSerializer::USE_ARRAYS_FOR_MAPS );
-
-		return new PropertySerializer( $fingerprintSerializer, $statementListSerializerMock );
+		return new PropertySerializer(
+			$termListSerializerMock,
+			$aliasGroupListSerializerMock,
+			$statementListSerializerMock,
+			$useObjectsForMaps
+		);
 	}
 
 	public function serializableProvider() {
@@ -85,6 +116,78 @@ class PropertySerializerTest extends SerializerBaseTest {
 		);
 
 		$property = Property::newFromType( '' );
+		$property->setId( 42 );
+		$provider[] = array(
+			array(
+				'type' => 'property',
+				'id' => 'P42',
+				'datatype' => '',
+				'claims' => array(),
+				'labels' => array(),
+				'descriptions' => array(),
+				'aliases' => array(),
+			),
+			$property
+		);
+
+		$property = Property::newFromType( '' );
+		$property->getFingerprint()->setLabel( 'en', 'foo' );
+		$provider[] = array(
+			array(
+				'type' => 'property',
+				'datatype' => '',
+				'claims' => array(),
+				'labels' => array(
+					'en' => array(
+						'lang' => 'en',
+						'value' => 'foo'
+					)
+				),
+				'descriptions' => array(),
+				'aliases' => array(),
+			),
+			$property
+		);
+
+		$property = Property::newFromType( '' );
+		$property->getFingerprint()->setDescription( 'en', 'foo' );
+		$provider[] = array(
+			array(
+				'type' => 'property',
+				'datatype' => '',
+				'claims' => array(),
+				'labels' => array(),
+				'descriptions' => array(
+					'en' => array(
+						'lang' => 'en',
+						'value' => 'foo'
+					)
+				),
+				'aliases' => array(),
+			),
+			$property
+		);
+
+		$property = Property::newFromType( '' );
+		$property->getFingerprint()->setAliasGroup( 'en', array( 'foo', 'bar' ) );
+		$provider[] = array(
+			array(
+				'type' => 'property',
+				'datatype' => '',
+				'claims' => array(),
+				'labels' => array(),
+				'descriptions' => array(),
+				'aliases' => array(
+					'en' => array(
+						'lang' => 'en',
+						'values' => array( 'foo', 'bar' )
+					)
+				),
+			),
+			$property
+		);
+
+		$property = Property::newFromType( '' );
 		$property->getStatements()->addNewStatement( new PropertyNoValueSnak( 42 ), null, null, 'test' );
 		$provider[] = array(
 			array(
@@ -110,6 +213,44 @@ class PropertySerializerTest extends SerializerBaseTest {
 		);
 
 		return $provider;
+	}
+
+	public function testPropertySerializerWithOptionObjectsForMaps() {
+		$serializer = $this->buildSerializer( true );
+
+		$property = Property::newFromType( '' );
+		$property->getFingerprint()->setLabel( 'en', 'foo' );
+		$property->getFingerprint()->setDescription( 'en', 'foo' );
+		$property->getFingerprint()->setAliasGroup( 'en', array( 'foo', 'bar' ) );
+
+		$labels = new \stdClass();
+		$labels->en = array(
+			'lang' => 'en',
+			'value' => 'foo'
+		);
+
+		$descriptions = new \stdClass();
+		$descriptions->en = array(
+			'lang' => 'en',
+			'value' => 'foo'
+		);
+
+		$aliases = new \stdClass();
+		$aliases->en = array(
+			'lang' => 'en',
+			'values' => array( 'foo', 'bar' )
+		);
+
+		$serial = array(
+			'type' => 'property',
+			'datatype' => '',
+			'labels' => $labels,
+			'descriptions' => $descriptions,
+			'aliases' => $aliases,
+			'claims' => array()
+		);
+
+		$this->assertEquals( $serial, $serializer->serialize( $property ) );
 	}
 
 }